Agents/External Agents/Web Scraper Agent/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "web-scraper-mcp"
version = "1.0.0"
description = "Headless browser web scraping MCP server for Solace Agent Mesh"
requires-python = ">=3.10"
dependencies = [
    "playwright>=1.49.0",
    "Pillow>=10.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.4",
]

[project.scripts]
web-scraper-mcp-server = "web_scraper_mcp.server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.ruff.lint.per-file-ignores]
# User-Agent strings are inherently long and must not be split
"src/web_scraper_mcp/config.py" = ["E501"]
